+++ package.json
@@ -2,10 +2,11 @@
   "name": "http-status-codes",
   "version": "1.0.0",
   "description": "HTTP Status Codes!",
-  "main": "dist/HTTPStatusCodes.js",
+  "main": "lib/status.ts",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc",
+    "dev": "tsc -w"
   },
   "author": "Shresht7",
   "license": "MIT",
